requirements.txt
azure-ai-documentintelligence==1.0.2
beautifulsoup4==4.13.4 
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langgraph==0.6.8
markdownify==1.2.0
msal==1.33.0
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
python-dotenv==1.1.1
pydantic==2.11.9
pydantic_settings==2.10.1
pyodbc==5.2.0
python-docx==1.2.0
ruff==0.13.1
tabulate==0.9.0
tiktoken>=0.9.0
weaviate-client==4.16.10
docx2pdf==0.1.8
fpdf==1.7.2
PyMuPDF==1.26.6

